ToDoList/src/main/java/Task/Task.java
package Task;

import java.time.LocalDateTime;

public class Task {
    private int id;
    private String name;
    private String content;
    LocalDateTime createDate;


    public Task (String name, String content, LocalDateTime createDate) {
        this.id = this.id + 1;
        this.name = name;
        this.content = content;
        this.createDate = createDate;
    }

    public int getId() {
        return id;
    }

    public String getName() {
        return name;
    }

    public String getContent() {
        return content;
    }

    public LocalDateTime getCreateDate() {
        return createDate;
    }


    public boolean isContentValid(String content) {
        return content.length() < 1000 && content.length() > 0;
    }

    public boolean isThirtyMinutesOver(LocalDateTime createDate) {
        if(this.createDate.getDayOfMonth() != createDate.getDayOfMonth()) return true;
        else if(this.createDate.getMinute() - createDate.getMinute() >= 30 && this.createDate.getHour() == createDate.getHour()) return true;
        else if(this.createDate.getMonth() != createDate.getMonth()) return true;
        else if(this.createDate.getMinute() <= createDate.getMinute() && this.createDate.getHour() != createDate.getHour())return true;
        else if(this.createDate.getMinute() >= createDate.getMinute() && this.createDate.getHour() != createDate.getHour() ){
            int i = 0;
            i = this.createDate.getMinute()-createDate.getMinute();
            i = 60 - i;
            if(i<30){
                return false;
            }else{
                return true;
            }
        }
        return false;
    }
    public boolean isTaskValid() {
        if(isThirtyMinutesOver(this.getCreateDate()) && isContentValid(this.getContent())) {
            return true;
        }
        return false;
    }
    public boolean EmailSenderService() {
        System.out.println("Il ne vous reste plus que 2 tâches!");
        return true;
    }
}
